feat: complete query migration to pure Rust core and fix all build errors
- Implement complete Rust query parser in /src/core/query_parser.rs
- Create FFI wrapper in /src/ffi/core/query_parser.rs
- Eliminate all 5 circular dependency calls (py.import -> Rust -> Python)
- Fix 48 compilation errors through systematic Result unwrapping
- Fix 14 PyO3 attribute errors by restructuring #[pymethods] blocks
- Achieve successful maturin develop --release build
- Update Python imports to match Rust exports
- Establish pure Rust core architecture for multi-language ports

Major architectural improvement: All query logic now resides in Rust core,
eliminating the problematic Python -> Rust -> Python circular dependency
anti-pattern and enabling clean multi-language port strategy.

+++ src/core/pool.rs
@@ -738,7 +738,7 @@ impl GraphPool {
         nodes.len().hash(&mut hasher);
         edges.len().hash(&mut hasher);
         subgraph_type.hash(&mut hasher);
-        let subgraph_id = hasher.finish() as crate::types::SubgraphId;
+        let subgraph_id = (hasher.finish() as usize) as crate::types::SubgraphId;
         Ok(subgraph_id)
     }
